+++ backend/de.metas.contracts/src/main/java/de/metas/contracts/interceptor/C_OrderLine.java
@@ -1,25 +1,31 @@
 package de.metas.contracts.interceptor;
 
-import static org.adempiere.model.InterfaceWrapperHelper.save;
-
+import de.metas.contracts.order.model.I_C_OrderLine;
+import de.metas.contracts.subscription.ISubscriptionBL;
+import de.metas.lang.SOTrx;
+import de.metas.order.IOrderLineBL;
+import de.metas.order.compensationGroup.GroupId;
+import de.metas.order.compensationGroup.OrderGroupCompensationChangesHandler;
+import de.metas.order.compensationGroup.OrderGroupRepository;
+import de.metas.util.Services;
+import lombok.NonNull;
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
 import org.adempiere.ad.persistence.ModelDynAttributeAccessor;
 import org.compiere.model.ModelValidator;
 import org.springframework.stereotype.Component;
 
-import de.metas.contracts.order.model.I_C_OrderLine;
-import de.metas.order.compensationGroup.GroupId;
-import de.metas.order.compensationGroup.OrderGroupCompensationChangesHandler;
-import de.metas.order.compensationGroup.OrderGroupRepository;
-import lombok.NonNull;
+import java.math.BigDecimal;
+
+import static org.adempiere.model.InterfaceWrapperHelper.save;
 
 @Interceptor(I_C_OrderLine.class)
 @Component
 public class C_OrderLine
 {
 	private static final ModelDynAttributeAccessor<I_C_OrderLine, Boolean> DYNATTR_SkipUpdatingGroupFlatrateConditions = new ModelDynAttributeAccessor<>("SkipUpdatingGroupFlatrateConditions", Boolean.class);
-
+	private final IOrderLineBL orderLineBL = Services.get(IOrderLineBL.class);
+	private final ISubscriptionBL subscriptionBL = Services.get(ISubscriptionBL.class);
 	private final OrderGroupCompensationChangesHandler groupChangesHandler;
 
 	public C_OrderLine(@NonNull final OrderGroupCompensationChangesHandler groupChangesHandler)
@@ -47,7 +53,7 @@ public class C_OrderLine
 	/**
 	 * In case the flatrate conditions for an order line is updated and that line is part of an compensation group,
 	 * then set the same flatrate conditions to all other lines from the same compensation group.
-	 * 
+	 *
 	 * @task https://github.com/metasfresh/metasfresh/issues/3150
 	 */
 	@ModelChange(timings = { ModelValidator.TYPE_AFTER_CHANGE }, ifColumnsChanged = I_C_OrderLine.COLUMNNAME_C_Flatrate_Conditions_ID, skipIfCopying = true)
@@ -94,5 +100,36 @@ public class C_OrderLine
 					DYNATTR_SkipUpdatingGroupFlatrateConditions.setValue(otherOrderLine, Boolean.TRUE);
 					save(otherOrderLine);
 				});
+	}
+
+	/**
+	 * Set QtyOrderedInPriceUOM, just to make sure is up2date.
+	 */
+	@ModelChange(timings = { ModelValidator.TYPE_BEFORE_NEW,
+			ModelValidator.TYPE_BEFORE_CHANGE
+	}, ifColumnsChanged = { de.metas.interfaces.I_C_OrderLine.COLUMNNAME_QtyEntered,
+			de.metas.interfaces.I_C_OrderLine.COLUMNNAME_Price_UOM_ID,
+			de.metas.interfaces.I_C_OrderLine.COLUMNNAME_C_UOM_ID,
+			de.metas.interfaces.I_C_OrderLine.COLUMNNAME_M_Product_ID
+	})
+	public void setQtyEnteredInPriceUOM(final I_C_OrderLine orderLine)
+	{
+		if (orderLine.getC_Flatrate_Conditions_ID() <= 0)
+		{
+			final BigDecimal qtyEnteredInPriceUOM = orderLineBL.convertQtyEnteredToPriceUOM(orderLine).toBigDecimal();
+			orderLine.setQtyEnteredInPriceUOM(qtyEnteredInPriceUOM);
+		}
+		else
+		{
+			final org.compiere.model.I_C_Order order = orderLine.getC_Order();
+			final SOTrx soTrx = SOTrx.ofBoolean(order.isSOTrx());
+
+			if (soTrx.isPurchase())
+			{
+				return; // leave this job to the adempiere standard callouts
+			}
+
+			subscriptionBL.updatePrices(orderLine, soTrx, true);
+		}
 	}
 }
