+++ backend/de.metas.contracts/src/main/java/de/metas/contracts/subscription/impl/SubscriptionBL.java
@@ -1,69 +1,13 @@
 package de.metas.contracts.subscription.impl;
 
-import static org.adempiere.model.InterfaceWrapperHelper.save;
-
-/*
- * #%L
- * de.metas.contracts
- * %%
- * Copyright (C) 2015 metas GmbH
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
- * published by the Free Software Foundation, either version 2 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public
- * License along with this program. If not, see
- * <http://www.gnu.org/licenses/gpl-2.0.html>.
- * #L%
- */
-
-import java.math.BigDecimal;
-import java.sql.Timestamp;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.GregorianCalendar;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Properties;
-import java.util.Set;
-
-import de.metas.common.util.time.SystemTime;
-import org.adempiere.ad.dao.IQueryBL;
-import org.adempiere.ad.trx.api.ITrx;
-import org.adempiere.ad.trx.api.ITrxManager;
-import org.adempiere.exceptions.AdempiereException;
-import org.adempiere.mm.attributes.api.IAttributeSetInstanceBL;
-import org.adempiere.model.InterfaceWrapperHelper;
-import org.adempiere.service.ISysConfigBL;
-import org.adempiere.util.lang.IAutoCloseable;
-import org.adempiere.util.lang.impl.TableRecordReference;
-import org.compiere.Adempiere;
-import org.compiere.model.I_AD_User;
-import org.compiere.model.I_C_BPartner;
-import org.compiere.model.MNote;
-import org.compiere.model.Query;
-import org.compiere.util.DB;
-import org.compiere.util.Env;
-import org.compiere.util.TimeUtil;
-import org.compiere.util.Trx;
-import org.compiere.util.TrxRunnable2;
-import org.slf4j.Logger;
-
 import com.google.common.base.Preconditions;
-
 import de.metas.adempiere.model.I_C_Order;
 import de.metas.bpartner.BPartnerContactId;
 import de.metas.bpartner.BPartnerId;
 import de.metas.bpartner.BPartnerLocationId;
 import de.metas.bpartner.service.BPartnerInfo;
 import de.metas.bpartner.service.IBPartnerOrgBL;
+import de.metas.common.util.time.SystemTime;
 import de.metas.contracts.Contracts_Constants;
 import de.metas.contracts.FlatrateTermPricing;
 import de.metas.contracts.IContractsDAO;
@@ -94,7 +38,9 @@ import de.metas.inoutcandidate.model.I_M_ShipmentSchedule;
 import de.metas.lang.SOTrx;
 import de.metas.logging.LogManager;
 import de.metas.monitoring.api.IMonitoringBL;
+import de.metas.order.IOrderLineBL;
 import de.metas.order.OrderId;
+import de.metas.order.OrderLinePriceUpdateRequest;
 import de.metas.ordercandidate.api.IOLCandBL;
 import de.metas.ordercandidate.api.IOLCandEffectiveValuesBL;
 import de.metas.pricing.IPricingResult;
@@ -106,12 +52,46 @@ import de.metas.product.IProductDAO;
 import de.metas.product.IProductPA;
 import de.metas.product.ProductAndCategoryId;
 import de.metas.product.ProductId;
+import de.metas.quantity.Quantity;
 import de.metas.tax.api.TaxCategoryId;
+import de.metas.uom.IUOMConversionBL;
 import de.metas.uom.UomId;
 import de.metas.util.Check;
 import de.metas.util.Services;
 import de.metas.workflow.api.IWFExecutionFactory;
 import lombok.NonNull;
+import org.adempiere.ad.dao.IQueryBL;
+import org.adempiere.ad.trx.api.ITrx;
+import org.adempiere.ad.trx.api.ITrxManager;
+import org.adempiere.exceptions.AdempiereException;
+import org.adempiere.mm.attributes.api.IAttributeSetInstanceBL;
+import org.adempiere.model.InterfaceWrapperHelper;
+import org.adempiere.service.ISysConfigBL;
+import org.adempiere.util.lang.IAutoCloseable;
+import org.adempiere.util.lang.impl.TableRecordReference;
+import org.compiere.Adempiere;
+import org.compiere.model.I_AD_User;
+import org.compiere.model.I_C_BPartner;
+import org.compiere.model.MNote;
+import org.compiere.model.Query;
+import org.compiere.util.DB;
+import org.compiere.util.Env;
+import org.compiere.util.TimeUtil;
+import org.compiere.util.Trx;
+import org.compiere.util.TrxRunnable2;
+import org.slf4j.Logger;
+
+import java.math.BigDecimal;
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.GregorianCalendar;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Properties;
+import java.util.Set;
+
+import static org.adempiere.model.InterfaceWrapperHelper.save;
 
 public class SubscriptionBL implements ISubscriptionBL
 {
@@ -132,7 +112,7 @@ public class SubscriptionBL implements ISubscriptionBL
 
 		newTerm.setC_OrderLine_Term_ID(ol.getC_OrderLine_ID());
 		newTerm.setC_Order_Term_ID(ol.getC_Order_ID());
-		
+
 		newTerm.setC_Flatrate_Conditions_ID(cond.getC_Flatrate_Conditions_ID());
 
 		// important: we need to use qtyEntered here, because qtyOrdered (which
@@ -597,7 +577,6 @@ public class SubscriptionBL implements ISubscriptionBL
 	}
 
 	/**
-	 *
 	 * @param term
 	 * @param currentDate
 	 * @return never returns {@code null}
@@ -1017,5 +996,80 @@ public class SubscriptionBL implements ISubscriptionBL
 				|| X_C_Flatrate_Term.CONTRACTSTATUS_EndingContract.equals(status);
 
 		return !isCancelledOrVoided;
+	}
+
+	@Override
+	public void updatePrices(
+			@NonNull final I_C_OrderLine ol,
+			@NonNull final SOTrx soTrx,
+			final boolean updatePriceEnteredAndDiscountOnlyIfNotAlreadySet)
+	{
+		final ISubscriptionBL subscriptionBL = Services.get(ISubscriptionBL.class);
+		final IUOMConversionBL uomConversionBL = Services.get(IUOMConversionBL.class);
+		final IOrderLineBL orderLineBL = Services.get(IOrderLineBL.class);
+		final IPriceListDAO priceListDAO = Services.get(IPriceListDAO.class);
+
+		final I_C_Flatrate_Conditions flatrateConditions = ol.getC_Flatrate_Conditions();
+		final org.compiere.model.I_C_Order order = ol.getC_Order();
+
+		final PricingSystemId pricingSysytemId;
+
+		if (flatrateConditions.getM_PricingSystem_ID() > 0)
+		{
+			pricingSysytemId = PricingSystemId.ofRepoId(flatrateConditions.getM_PricingSystem_ID());
+		}
+		else
+		{
+			pricingSysytemId = PricingSystemId.ofRepoIdOrNull(order.getM_PricingSystem_ID());
+		}
+
+		final BPartnerLocationId bpLocationId = BPartnerLocationId.ofRepoId(ol.getC_BPartner_ID(), ol.getC_BPartner_Location_ID());
+		final Timestamp date = order.getDateOrdered();
+
+		final PriceListId subscriptionPLId = priceListDAO.retrievePriceListIdByPricingSyst(pricingSysytemId, bpLocationId, soTrx);
+
+		final int numberOfRuns = subscriptionBL.computeNumberOfRuns(flatrateConditions.getC_Flatrate_Transition(), date);
+
+		final Properties ctx = Env.getCtx();
+		final ProductId productId = ProductId.ofRepoIdOrNull(ol.getM_Product_ID());
+		final ProductAndCategoryId productAndCategoryId = Services.get(IProductDAO.class).retrieveProductAndCategoryIdByProductId(productId);
+		final I_C_Flatrate_Matching matching = subscriptionBL.retrieveMatching(
+				ctx,
+				ol.getC_Flatrate_Conditions_ID(),
+				productAndCategoryId,
+				ITrx.TRXNAME_None);
+
+		final Quantity qtyEntered = orderLineBL.getQtyEntered(ol);
+		final Quantity qtyOrdered = uomConversionBL.convertToProductUOM(qtyEntered, productId);
+
+		final Quantity qtyOrderedPerRun;
+		if (matching != null && matching.getQtyPerDelivery().signum() > 0)
+		{
+			final Quantity qtyPerDelivery = Quantity.of(matching.getQtyPerDelivery(), qtyOrdered.getUOM());
+			qtyOrderedPerRun = qtyPerDelivery.min(qtyOrdered);
+		}
+		else
+		{
+			qtyOrderedPerRun = qtyOrdered;
+		}
+
+		// priceQty is the qty do be delivered during one complete subscription term
+		final Quantity priceQty = qtyOrderedPerRun.multiply(numberOfRuns);
+
+		// qty ordered needs to be set because it will be used to compute the
+		// line's NetLineAmount in MOrderLine.beforeSave()
+		ol.setQtyOrdered(priceQty.toBigDecimal());
+
+		ol.setQtyEnteredInPriceUOM(priceQty.toBigDecimal());
+
+		// now compute the new prices
+		orderLineBL.updatePrices(OrderLinePriceUpdateRequest.builder()
+				.orderLine(ol)
+				.priceListIdOverride(subscriptionPLId)
+				.qtyOverride(priceQty)
+				.resultUOM(OrderLinePriceUpdateRequest.ResultUOM.PRICE_UOM)
+				.updatePriceEnteredAndDiscountOnlyIfNotAlreadySet(updatePriceEnteredAndDiscountOnlyIfNotAlreadySet)
+				.updateLineNetAmt(true)
+				.build());
 	}
 }
